+++ feature/mypage/src/main/java/com/teamoffroad/feature/mypage/navigation/MypageNavigation.kt
@@ -5,7 +5,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.teamoffroad.core.navigation.MainTabRoute
-import com.teamoffroad.feature.mypage.MypageScreen
+import com.teamoffroad.feature.mypage.presentation.MypageScreen
 
 fun NavController.navigateToMyPage(navOptions: NavOptions) {
     navigate(MainTabRoute.MyPage, navOptions)
